instantiate by lazy SiopOpenId4Vp in order to get the overriden ktorHttpClientFactory; move Signer tests to correct package
