+++ wallet-core/src/androidTest/java/eu/europa/ec/eudi/wallet/issue/openid4vci/JWSProofSignerTest.kt
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package eu.europa.ec.eudi.wallet.document.issue.opeid4vci
+package eu.europa.ec.eudi.wallet.issue.openid4vci
 
 import android.app.KeyguardManager
 import android.content.Context
@@ -29,8 +29,6 @@ import eu.europa.ec.eudi.openid4vci.CNonce
 import eu.europa.ec.eudi.openid4vci.JwtBindingKey
 import eu.europa.ec.eudi.wallet.document.CreateDocumentResult
 import eu.europa.ec.eudi.wallet.document.DocumentManager
-import eu.europa.ec.eudi.wallet.issue.openid4vci.JWSProofSigner
-import eu.europa.ec.eudi.wallet.issue.openid4vci.ProofSigner
 import org.junit.Assert
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
